can't check if an unsigned type is negative, used signed.
after discussiong with thorduri

+++ sys/dev/pci/drm/drm_irq.c
@@ -141,7 +141,7 @@ struct timeval
 ns_to_timeval(const int64_t nsec)
 {
 	struct timeval tv;
-	uint32_t rem;
+	int64_t rem;
 
 	if (nsec == 0) {
 		tv.tv_sec = 0;
